+++ tools/lsp/parser/parser/parser.go
@@ -46,6 +46,7 @@ func init() {
 	topLevelParsers = map[token.TokenType]defParser{
 		token.WORKFLOW: parseWorkflowDef,
 		token.ACTIVITY: parseActivityDef,
+		token.WORKER:   parseWorkerDef,
 	}
 
 	workflowStmtParsers = map[token.TokenType]stmtParser{